+++ test/EntityDb.Common.Tests/Implementations/Tags/CountTag.cs
@@ -2,7 +2,7 @@ using EntityDb.Abstractions.Tags;
 
 namespace EntityDb.Common.Tests.Implementations.Tags;
 
-public record CountTag(int Number) : ITag
+public record CountTag(ulong Number) : ITag
 {
     public string Label => $"{Number}";
     public string Value => "";
